+++ Tests/Editor/RecommendationTests.cs
@@ -291,7 +291,7 @@ namespace Unity.MultiplayerCenterTests
             };
             var expectedPackages = new List<string> {"com.unity.netcode", "com.unity.services.multiplayer", 
                 "com.unity.netcode.gameobjects", // added as additional package
-                "com.unity.multiplayer.center.quickstart"}; // always added
+                "com.unity.multiplayer.center.quickstart@1.0.0-pre.1"}; // always added
 
             RecommendationUtils.GetPackagesWithAdditionalPackages(mainPackages, out var allPackages, out var allNames, out var tooltip);
             expectedPackages.Sort();